tests/wpt/tests/layout-instability/content-visibility-auto-onscreen.html

<!DOCTYPE html>
<title>Layout Instability: on-screen content-visibility:auto content</title>
<link rel="help" href="https://wicg.github.io/layout-instability/" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/util.js"></script>
<script>
// These scripts need to be before the contents because we need to ensure no
// layout shifts during page load.
promise_test(async () => {
  const watcher = new ScoreWatcher;

  // Wait for the initial render to complete.
  await waitForAnimationFrames(2);
  assert_equals(watcher.score, 0);
}, 'on-screen content-visibility:auto');
</script>
<style>
  #target {
    content-visibility: auto;
    contain-intrinsic-size: 1px;
    width: 100px;
  }
</style>
<div id=target>
  <div style="width: 100px; height: 100px; background: blue"></div>
</div>
